refactor: CarService 제거
