+++ src/main/java/Main.java
@@ -1,12 +1,10 @@
 import controller.CarController;
-import service.CarService;
 import utils.RacingRandomNumberGenerator;
 
 public class Main {
 
     public static void main(String[] args) {
-        CarService carService = new CarService();
-        CarController carController = new CarController(carService);
+        CarController carController = new CarController();
 
         carController.run(new RacingRandomNumberGenerator());
     }
